etl/requirements.txt
loguru==0.5.3
psycopg2-binary==2.9.1
elasticsearch==7.13.1
backoff==1.11.1
redis==3.5.3